Connecting front-end with the back-end

+++ Flask-Web-App-Tutorial-main/website/views.py
@@ -6,32 +6,44 @@ import json
 
 views = Blueprint('views', __name__)
 
-
-@views.route('/', methods=['GET', 'POST'])
+# 🌱 Your existing note-taking home
+@views.route('/notes', methods=['GET', 'POST'])
 @login_required
-def home():
+def notes():
     if request.method == 'POST': 
-        note = request.form.get('note')#Gets the note from the HTML 
-
+        note = request.form.get('note')
         if len(note) < 1:
             flash('Note is too short!', category='error') 
         else:
-            new_note = Note(data=note, user_id=current_user.id)  #providing the schema for the note 
-            db.session.add(new_note) #adding the note to the database 
+            new_note = Note(data=note, user_id=current_user.id)
+            db.session.add(new_note)
             db.session.commit()
             flash('Note added!', category='success')
 
     return render_template("home.html", user=current_user)
 
-
 @views.route('/delete-note', methods=['POST'])
-def delete_note():  
-    note = json.loads(request.data) # this function expects a JSON from the INDEX.js file 
+@login_required
+def delete_note():
+    note = json.loads(request.data)
     noteId = note['noteId']
     note = Note.query.get(noteId)
-    if note:
-        if note.user_id == current_user.id:
-            db.session.delete(note)
-            db.session.commit()
+    if note and note.user_id == current_user.id:
+        db.session.delete(note)
+        db.session.commit()
 
     return jsonify({})
+
+# 🌍 New EcoLink pages
+@views.route('/')
+def home():
+    return render_template("index.html", user=current_user)
+
+@views.route('/about-us')
+def about_us():
+    return render_template("about_us.html", user=current_user)
+
+@views.route('/food-forests')
+def food_forests():
+    return render_template("food_forests.html", user=current_user)
+
